+++ crates/voicevox_core/src/engine/talk/audio_query/validated.rs
@@ -135,17 +135,24 @@ impl AudioQuery {
 }
 
 pub trait Validate: DeserializeOwned {
+    const NAME: &str;
     fn validate(&self) -> crate::Result<()>;
+
+    fn validation_error_description() -> String {
+        format!("不正な{}です", Self::NAME)
+    }
 }
 
 #[duplicate_item(
-    T validation;
-    [ AudioQuery ] [ Self::validate ];
-    [ AccentPhrase ] [ Self::validate ];
-    [ Mora ] [ Self::validate ];
-    [ Vec<AccentPhrase> ] [ |this: &Self| this.iter().try_for_each(AccentPhrase::validate) ];
+    T S validation;
+    [ AudioQuery ] [ "AudioQuery" ] [ Self::validate ];
+    [ AccentPhrase ] [ "アクセント句" ] [ Self::validate ];
+    [ Mora ] [ "モーラ" ] [ Self::validate ];
+    [ Vec<AccentPhrase> ] [ "アクセント句の列" ] [ |this: &Self| this.iter().try_for_each(AccentPhrase::validate) ];
 )]
 impl Validate for T {
+    const NAME: &str = S;
+
     fn validate(&self) -> crate::Result<()> {
         (validation)(self)
     }
@@ -243,7 +250,7 @@ impl<'original> ValidatedMora<'original> {
 
         fn error(source: InvalidQueryErrorSource) -> InvalidQueryError {
             InvalidQueryError {
-                what: "モーラ",
+                what: Mora::NAME,
                 value: None,
                 source: Some(source),
             }
@@ -397,7 +404,7 @@ impl<'original> ValidatedAccentPhrase<'original> {
 
         fn error(source: InvalidQueryErrorSource) -> InvalidQueryError {
             InvalidQueryError {
-                what: "アクセント句",
+                what: AccentPhrase::NAME,
                 value: None,
                 source: Some(source),
             }
@@ -528,7 +535,7 @@ impl<'original> ValidatedAudioQuery<'original> {
 
         fn error(source: InvalidQueryErrorSource) -> InvalidQueryError {
             InvalidQueryError {
-                what: "AudioQuery",
+                what: AudioQuery::NAME,
                 value: None,
                 source: Some(source),
             }