+++ crates/voicevox_core_java_api/lib/src/main/java/jp/hiroshiba/voicevoxcore/blocking/Synthesizer.java
@@ -14,6 +14,7 @@ import jp.hiroshiba.voicevoxcore.CharacterMeta;
 import jp.hiroshiba.voicevoxcore.StyleType;
 import jp.hiroshiba.voicevoxcore.exceptions.InvalidModelDataException;
 import jp.hiroshiba.voicevoxcore.exceptions.RunModelException;
+import jp.hiroshiba.voicevoxcore.internal.Convert;
 import jp.hiroshiba.voicevoxcore.internal.Dll;
 
 /**
@@ -225,7 +226,7 @@ public class Synthesizer {
     if (!Utils.isU32(styleId)) {
       throw new IllegalArgumentException("styleId");
     }
-    String accentPhrasesJson = new Gson().toJson(accentPhrases);
+    String accentPhrasesJson = Convert.jsonFromQueryLike(accentPhrases, "アクセント句の列");
     String replacedAccentPhrasesJson = rsReplaceMoraData(accentPhrasesJson, styleId, false);
     return new ArrayList<>(
         Arrays.asList(new Gson().fromJson(replacedAccentPhrasesJson, AccentPhrase[].class)));
@@ -245,7 +246,7 @@ public class Synthesizer {
     if (!Utils.isU32(styleId)) {
       throw new IllegalArgumentException("styleId");
     }
-    String accentPhrasesJson = new Gson().toJson(accentPhrases);
+    String accentPhrasesJson = Convert.jsonFromQueryLike(accentPhrases, "アクセント句の列");
     String replacedAccentPhrasesJson = rsReplacePhonemeLength(accentPhrasesJson, styleId, false);
     return new ArrayList<>(
         Arrays.asList(new Gson().fromJson(replacedAccentPhrasesJson, AccentPhrase[].class)));
@@ -265,7 +266,7 @@ public class Synthesizer {
     if (!Utils.isU32(styleId)) {
       throw new IllegalArgumentException("styleId");
     }
-    String accentPhrasesJson = new Gson().toJson(accentPhrases);
+    String accentPhrasesJson = Convert.jsonFromQueryLike(accentPhrases, "アクセント句の列");
     String replacedAccentPhrasesJson = rsReplaceMoraPitch(accentPhrasesJson, styleId, false);
     return new ArrayList<>(
         Arrays.asList(new Gson().fromJson(replacedAccentPhrasesJson, AccentPhrase[].class)));
@@ -465,8 +466,7 @@ public class Synthesizer {
       if (!Utils.isU32(styleId)) {
         throw new IllegalArgumentException("styleId");
       }
-      Gson gson = new Gson();
-      String queryJson = gson.toJson(this.audioQuery);
+      String queryJson = Convert.jsonFromQueryLike(this.audioQuery, "AudioQuery");
       return synthesizer.rsSynthesis(queryJson, this.styleId, this.interrogativeUpspeak);
     }
   }